relocate gh types
