+++ openhands/server/routes/github.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends
 from fastapi.responses import JSONResponse
 
 from openhands.server.auth import get_user_id
-from openhands.server.services.gh_types import GitHubRepository, GitHubUser
+from openhands.server.data_models.gh_types import GitHubRepository, GitHubUser
 from openhands.server.services.github_service import GitHubService
 from openhands.server.shared import server_config
 from openhands.server.types import GhAuthenticationError, GHUnknownException
